Update nep-dynamic-invoke.mediawiki
This is really great! I did a quick edit pass to fill in some sharding details. I also think `deterministic` is not the right word to be using here, as execution is deterministic either way, with dynamic calls or not. The problem is how much static analysis we can do in advance, to shard resources appropriately.