NEP Dynamic Invoke (#19)
* initial nep-9 work

* Update README.mediawiki

Fix layout

* removed nep version number

* Update nep-dynamic-invoke.mediawiki

This is really great! I did a quick edit pass to fill in some sharding details. I also think `deterministic` is not the right word to be using here, as execution is deterministic either way, with dynamic calls or not. The problem is how much static analysis we can do in advance, to shard resources appropriately.

* added sample smart contract, moved some considerations to the rationale section

* formatting fixes

* formatting fixes

* formatting change

* added links, copy edits

* copy edits

* copy edits

* copy changes

* copy changes

* copy edit.

* Added proposed fees

Added sample implementation for an updated fee schedule.

* Updated fees for creation of Smart Contract

* updated to be nep-4

* change state to final

* formatting