Update send input args to be more verbose
There is a small race condition for the connected account, as all dapp to wallet provider communications are assumed to be async. There is a chance that the dapp developer submits the send request to the wallet provider as the `ACCOUNT_CHANGED` event is fired, resulting in the dapp developer sending assets from the incorrect account.